Bug #841311 modify the 8 bits backlight value  to 16 bits
[root cause  ] modify the 8 bits backlight value  to 16 bits
[changes     ] PQTuningParmDefine.h
[side effects] no
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:no
[board]:sp9863a1h10
[test case]:pq tool case
[reviewers   ] tim.xia
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

+++ PQTune/PQTuningParmDefine.h
@@ -75,8 +75,8 @@ struct slp_cfg
 
 struct slp_mapping_table_item
 {
-    u8 ambient;
-    u8 slp_brightness_factor;
+    u16 ambient;
+    u16 slp_brightness_factor;
 };
 
 struct slp_mapping_table
@@ -87,8 +87,8 @@ struct slp_mapping_table
 
 struct bl_mapping_table_item
 {
-    u8 ambient;
-    u8 backlight;
+    u16 ambient;
+    u16 backlight;
 };
 
 struct bl_mapping_table
